+++ functional_tests/test.sh
@@ -1,0 +1,7 @@
+cd "$(dirname "$0")"
+
+cd ..
+cargo build -r
+cd -
+
+./test_parenthesis/test.sh